add one hot cpu kernel;
